Move some functions into logic-apps-shared
