+++ libs/designer-ui/src/lib/connectors/__test__/predicates.spec.tsx
@@ -1,12 +1,12 @@
-import type { Connector, OperationApi } from '../../../models';
-import { isBuiltInConnector } from '../index';
 import * as DesignerClientServices from '@microsoft/designer-client-services-logic-apps';
+import type { Connector, OperationApi } from '@microsoft/logic-apps-shared';
+import { isBuiltInConnector } from '../index';
 
 const getMinimalHostService = (): DesignerClientServices.IHostService => ({
   fetchAndDisplayContent: jest.fn(),
 });
 
-describe('lib/connectors/connectorProperties', () => {
+describe('lib/connectors', () => {
   afterEach(() => {
     jest.restoreAllMocks();
   });
