+++ libs/designer-ui/src/lib/connectors/index.ts
@@ -1,0 +1,1 @@
+export * from './predicates';
