+++ libs/designer-ui/src/lib/connectors/predicates.ts
@@ -1,8 +1,6 @@
-import type { Connector, OperationApi } from '../../models';
-import { isBuiltInConnectorId, isCustomConnectorId } from '../connections';
-import { isString } from '../functions';
-import { getAllConnectorProperties } from './connectorProperties';
 import { HostService } from '@microsoft/designer-client-services-logic-apps';
+import type { Connector, OperationApi} from '@microsoft/logic-apps-shared';
+import { getAllConnectorProperties, isBuiltInConnectorId, isCustomConnectorId, isString } from '@microsoft/logic-apps-shared';
 
 export const isBuiltInConnector = (connector: Connector | OperationApi | string): boolean => {
   const hostIsBuiltInConnectorFn = HostService()?.isBuiltInConnector;
