+++ libs/designer-ui/src/lib/index.ts
@@ -19,6 +19,7 @@ export * from './checkbox';
 // export * from './colorizer'
 export * from './code';
 export * from './combobox';
+export * from './connectors';
 export { default as UIConstants } from './constants';
 export * from './html';
 // export * from './conditioncontrol';
