+++ libs/designer-ui/src/lib/panel/recommendationpanel/searchResult.tsx
@@ -1,16 +1,16 @@
+import { List, Text } from '@fluentui/react';
+import { Spinner } from '@fluentui/react-components';
+import type { DiscoveryOpArray, DiscoveryOperation, DiscoveryResultTypes } from '@microsoft/logic-apps-shared';
+import type { PropsWithChildren } from 'react';
+import React, { useMemo } from 'react';
+import { useIntl } from 'react-intl';
 import NoResultsSvg from '../../../assets/search/noResults.svg';
 import { AriaSearchResultsAlert } from '../../ariaSearchResults/ariaSearchResultsAlert';
+import { isBuiltInConnector } from '../../connectors';
 import { getConnectorCategoryString } from '../../utils';
 import type { OperationActionData } from './interfaces';
 import { OperationSearchCard } from './operationSearchCard';
 import { OperationSearchGroup } from './operationSearchGroup';
-import { List, Text } from '@fluentui/react';
-import { Spinner } from '@fluentui/react-components';
-import type { DiscoveryOpArray, DiscoveryOperation, DiscoveryResultTypes } from '@microsoft/logic-apps-shared';
-import { isBuiltInConnector } from '@microsoft/logic-apps-shared';
-import type { PropsWithChildren } from 'react';
-import React, { useMemo } from 'react';
-import { useIntl } from 'react-intl';
 
 export type SearchResultsGridProps = {
   isLoadingMore: boolean;
