+++ libs/designer-ui/src/lib/utils/utils.ts
@@ -1,6 +1,7 @@
-import Constants from '../constants';
 import type { Connector, OperationApi } from '@microsoft/logic-apps-shared';
-import { equals, getIntl, isBuiltInConnector, isCustomConnector } from '@microsoft/logic-apps-shared';
+import { equals, getIntl } from '@microsoft/logic-apps-shared';
+import { isBuiltInConnector, isCustomConnector } from '../connectors';
+import Constants from '../constants';
 
 /**
  * Returns a string with a duration, possibly abbreviated, e.g., 15s or 15 second(s)
