--- libs/logic-apps-shared/src/utils/src/lib/helpers/connectors/index.ts
@@ -1,2 +1,0 @@
-export * from './connectorPredicates';
-export * from './connectorProperties';
